refactor: 패키지 이름 수정 및 자동정렬을 통한 품질 개선 Shop도메인 패키지 이름 수정 DTO->Dto api -> presentation 자동정렬 단축키 사용
