+++ src/main/java/org/hmanwon/domain/community/board/dao/BoardRepository.java
@@ -1,0 +1,7 @@
+package org.hmanwon.domain.community.board.dao;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BoardRepository extends JpaRepository {
+
+}
