--- src/main/java/org/hmanwon/domain/community/board/repository/BoardRepository.java
@@ -1,7 +1,0 @@
-package org.hmanwon.domain.community.board.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BoardRepository extends JpaRepository {
-
-}
