+++ src/main/java/org/hmanwon/domain/community/comment/application/CommentService.java
@@ -3,9 +3,9 @@ package org.hmanwon.domain.community.comment.application;
 import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.hmanwon.domain.community.board.entity.Board;
-import org.hmanwon.domain.community.board.repository.BoardRepository;
-import org.hmanwon.domain.community.comment.dto.request.CommentRequestDTO;
-import org.hmanwon.domain.community.comment.dto.response.CommentResponseDTO;
+import org.hmanwon.domain.community.board.dao.BoardRepository;
+import org.hmanwon.domain.community.comment.dto.request.CommentRequestDto;
+import org.hmanwon.domain.community.comment.dto.response.CommentResponseDto;
 import org.hmanwon.domain.community.comment.entity.Comment;
 import org.hmanwon.domain.community.comment.dao.CommentRepository;
 import org.hmanwon.domain.member.dao.MemberRepository;
@@ -27,7 +27,7 @@ public class CommentService {
      * @param commentRequestDTO 댓글 요청 DTO
      * @return 댓글 응답 DTO
      */
-    public CommentResponseDTO createComment(Long memberId, CommentRequestDTO commentRequestDTO) {
+    public CommentResponseDto createComment(Long memberId, CommentRequestDto commentRequestDTO) {
         Comment comment = Comment.builder()
             .board((Board) boardRepository.findById(commentRequestDTO.boardId()).orElseThrow()) //추후 변경 예정
             .member((Member) memberRepository.findById(memberId).orElseThrow()) //추후 변경 예정
@@ -36,7 +36,7 @@ public class CommentService {
 
         commentRepository.save(comment);
 
-        return CommentResponseDTO.entityFromDTO(comment);
+        return CommentResponseDto.entityFromDTO(comment);
     }
 
     /***
@@ -45,10 +45,10 @@ public class CommentService {
      * @param commentId 댓글 ID
      * @return 댓글
      */
-    public CommentResponseDTO deleteCommentById(Long memberId, Long commentId) {
+    public CommentResponseDto deleteCommentById(Long memberId, Long commentId) {
         Comment comment = (Comment) commentRepository.findById(commentId).orElseThrow();
         commentRepository.delete(comment);
 
-        return CommentResponseDTO.entityFromDTO(comment);
+        return CommentResponseDto.entityFromDTO(comment);
     }
 }
