--- src/main/java/org/hmanwon/domain/community/comment/dto/request/CommentRequestDTO.java
@@ -1,12 +1,0 @@
-package org.hmanwon.domain.community.comment.dto.request;
-
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.NotNull;
-
-public record CommentRequestDTO(
-    @NotNull
-    Long boardId,
-    @NotBlank
-    String content
-) {
-}
