+++ src/main/java/org/hmanwon/domain/community/comment/dto/request/CommentRequestDto.java
@@ -1,0 +1,12 @@
+package org.hmanwon.domain.community.comment.dto.request;
+
+import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.NotNull;
+
+public record CommentRequestDto(
+    @NotNull
+    Long boardId,
+    @NotBlank
+    String content
+) {
+}
