--- src/main/java/org/hmanwon/domain/display/dao/ItemRepository.java
@@ -1,10 +1,0 @@
-package org.hmanwon.domain.display.dao;
-
-import org.hmanwon.domain.display.entity.Menu;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ItemRepository extends JpaRepository<Menu, Long> {
-
-}
