--- src/main/java/org/hmanwon/domain/display/exception/DisplayException.java
@@ -1,11 +1,0 @@
-package org.hmanwon.domain.display.exception;
-
-import org.hmanwon.global.common.exception.DefaultException;
-import org.hmanwon.global.common.exception.ExceptionCode;
-
-public class DisplayException extends DefaultException {
-
-    public DisplayException(ExceptionCode exceptionCode) {
-        super(exceptionCode);
-    }
-}
