--- src/main/java/org/hmanwon/domain/display/init/dto/FirstJsonReadDTO.java
@@ -1,10 +1,0 @@
-package org.hmanwon.domain.display.init.dto;
-
-import java.util.List;
-import lombok.Getter;
-
-@Getter
-public class FirstJsonReadDTO {
-    private String description;
-    private List<SeoulGoodShopDTO> DATA;
-}
