--- src/main/java/org/hmanwon/domain/display/init/dto/ItemDTO.java
@@ -1,6 +1,0 @@
-package org.hmanwon.domain.display.init.dto;
-
-public class ItemDTO {
-    private String itemName;
-    private Integer itemPrice;
-}
