--- src/main/java/org/hmanwon/domain/display/init/exception/InitDisplayException.java
@@ -1,11 +1,0 @@
-package org.hmanwon.domain.display.init.exception;
-
-import org.hmanwon.global.common.exception.DefaultException;
-import org.hmanwon.global.common.exception.ExceptionCode;
-
-public class InitDisplayException extends DefaultException {
-
-    public InitDisplayException(ExceptionCode exceptionCode) {
-        super(exceptionCode);
-    }
-}
