+++ src/main/java/org/hmanwon/domain/shop/dao/MenuRepository.java
@@ -1,0 +1,10 @@
+package org.hmanwon.domain.shop.dao;
+
+import org.hmanwon.domain.shop.entity.Menu;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MenuRepository extends JpaRepository<Menu, Long> {
+
+}
