+++ src/main/java/org/hmanwon/domain/shop/exception/ShopException.java
@@ -1,0 +1,11 @@
+package org.hmanwon.domain.shop.exception;
+
+import org.hmanwon.global.common.exception.DefaultException;
+import org.hmanwon.global.common.exception.ExceptionCode;
+
+public class ShopException extends DefaultException {
+
+    public ShopException(ExceptionCode exceptionCode) {
+        super(exceptionCode);
+    }
+}
