+++ src/main/java/org/hmanwon/domain/shop/exception/ShopExceptionCode.java
@@ -1,0 +1,19 @@
+package org.hmanwon.domain.shop.exception;
+
+import static org.springframework.http.HttpStatus.INTERNAL_SERVER_ERROR;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.hmanwon.global.common.exception.ExceptionCode;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@AllArgsConstructor
+public enum ShopExceptionCode implements ExceptionCode {
+
+    NOT_FOUND_FILE(INTERNAL_SERVER_ERROR, "Display Error 01", "파일이 존재하지 않습니다.");
+
+    private final HttpStatus status;
+    private final String code;
+    private final String msg;
+}
