+++ src/main/java/org/hmanwon/domain/shop/init/dto/FirstJsonReadDTO.java
@@ -1,0 +1,10 @@
+package org.hmanwon.domain.shop.init.dto;
+
+import java.util.List;
+import lombok.Getter;
+
+@Getter
+public class FirstJsonReadDTO {
+    private String description;
+    private List<SeoulGoodShopDTO> DATA;
+}
