+++ src/main/java/org/hmanwon/domain/shop/init/dto/ItemDTO.java
@@ -1,0 +1,6 @@
+package org.hmanwon.domain.shop.init.dto;
+
+public class ItemDTO {
+    private String itemName;
+    private Integer itemPrice;
+}
