+++ src/main/java/org/hmanwon/global/common/exception/DefaultExceptionCode.java
@@ -9,7 +9,8 @@ import org.springframework.http.HttpStatus;
 public enum DefaultExceptionCode implements ExceptionCode {
 
 
-    INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "INTERNAL_SERVER_ERROR","서버에 오류가 발생했습니다."),
+    INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "INTERNAL_SERVER_ERROR",
+        "서버에 오류가 발생했습니다."),
     BAD_REQUEST(HttpStatus.BAD_REQUEST, "BAD_REQUEST", "잘못된 요청 입니다."),
 
     ;
