+++ src/main/java/org/hmanwon/global/common/exception/ExceptionCode.java
@@ -4,7 +4,10 @@ import org.springframework.http.HttpStatus;
 
 
 public interface ExceptionCode {
+
     HttpStatus getStatus();
+
     String getCode();
+
     String getMsg();
 }
